requirements.txt
aiogram>=3.0.0
sqlalchemy>=2.0.0
pandas>=1.3.0
matplotlib>=3.5.0
python-dotenv>=0.19.0
xlsxwriter>=3.0.0
xhtml2pdf>=0.2.8
